% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{$.enum}
\alias{$.enum}
\title{Get value from enum object}
\usage{
\method{$}{enum}(x, name)
}
\arguments{
\item{x}{An enum object}

\item{name}{The name of the field to access (should be "value")}
}
\value{
The value of the enum object
}
\description{
Retrieves the value of an enum object.
}
