% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{$.interface_object}
\alias{$.interface_object}
\title{Get a property from an interface object}
\usage{
\method{$}{interface_object}(x, name)
}
\arguments{
\item{x}{An interface object}

\item{name}{The name of the property to get}
}
\description{
Get a property from an interface object
}
