% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement.R
\name{$.validated_list}
\alias{$.validated_list}
\title{Custom accessor for validated_list objects}
\usage{
\method{$}{validated_list}(x, i)
}
\arguments{
\item{x}{The validated_list object}

\item{i}{The name of the property to access}
}
\value{
The value of the property
}
\description{
Custom accessor for validated_list objects
}
