% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{$<-.interface_object}
\alias{$<-.interface_object}
\title{Set a property in an interface object}
\usage{
\method{$}{interface_object}(x, name) <- value
}
\arguments{
\item{x}{An interface object}

\item{name}{The name of the property to set}

\item{value}{The new value for the property}
}
\value{
The modified interface object, invisibly.
}
\description{
Set a property in an interface object
}
