% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.frame.R
\name{$<-.typed_frame}
\alias{$<-.typed_frame}
\title{Modify a typed data frame using $}
\usage{
\method{$}{typed_frame}(x, col_name) <- value
}
\arguments{
\item{x}{A typed data frame.}

\item{col_name}{The name of the column to modify or add.}

\item{value}{The new value to assign.}
}
\value{
The modified typed data frame.
}
\description{
Allows modifying a typed data frame using the $ operator, with validation checks.
}
