% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement.R
\name{$<-.validated_list}
\alias{$<-.validated_list}
\title{Custom assignment for validated_list objects}
\usage{
\method{$}{validated_list}(x, i) <- value
}
\arguments{
\item{x}{The validated_list object}

\item{i}{The name of the property to assign}

\item{value}{The value to assign}
}
\value{
The updated validated_list object
}
\description{
Custom assignment for validated_list objects
}
