% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_interface}
\alias{check_interface}
\title{Check if a value implements an interface}
\usage{
check_interface(value, interface)
}
\arguments{
\item{value}{The value to check}

\item{interface}{The interface to check against}
}
\value{
Logical indicating whether the value implements the interface
}
\description{
Check if a value implements an interface
}
\keyword{internal}
