% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_priors.R
\docType{data}
\name{gene_priors}
\alias{gene_priors}
\title{Gene priors for simulated data.}
\format{
A 39-vector.
}
\usage{
gene_priors
}
\value{
A 39-vector.
}
\description{
A numeric 39-vector containing a prior probability for each gene.
This data is included to demonstrate the usage of the interface function.
The priors can be generated using the interface_priors function.
}
\examples{
data(gene_priors)
}
\keyword{datasets}
