% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcp.R
\docType{data}
\name{grcp}
\alias{grcp}
\title{Gene-level colocalization probabilities for simulated data.}
\format{
A 39-vector.
}
\usage{
grcp
}
\value{
A 39-vector.
}
\description{
A numeric 39-vector containing gene-level colocalization probabilities
generated from the fastENLOC method. This data is included to demonstrate the
usage of the interface_priors function.
}
\examples{
data(grcp)
}
\keyword{datasets}
