% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.frame.R
\name{handle_violation}
\alias{handle_violation}
\title{Handle violations based on the specified action}
\usage{
handle_violation(message, action)
}
\arguments{
\item{message}{The error message to be handled.}

\item{action}{The action to take: "error", "warning", or "silent".}
}
\description{
Handles violations by either throwing an error, issuing a warning, or doing nothing, depending on the specified action.
}
