% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement.R
\name{implement}
\alias{implement}
\title{Implement an Interface}
\usage{
implement(interface, ..., validate_on_access = NULL)
}
\arguments{
\item{interface}{An Interface object}

\item{...}{Properties to implement the interface}

\item{validate_on_access}{Logical, whether to validate on access}
}
\value{
An object implementing the interface
}
\description{
Implement an Interface
}
