% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_priors.R
\name{interface_priors}
\alias{interface_priors}
\title{Estimate gene priors for causal gene fine-mapping.}
\usage{
interface_priors(GRCP_vec, p12, TWAS_pi, n_cis)
}
\arguments{
\item{GRCP_vec}{A vector of colocalization probabilities for the genes of
interest}

\item{p12}{A prior enrichment parameter from colocalization analysis.
Represents the prior probability that a variant in the region is both a
causal GWAS variant and a causal eQTL.}

\item{n_cis}{A vector containing the number of cis variants for each gene
of interest. Order should match that of GRCP_vec.}

\item{TWAS_z}{A vector of TWAS z-scores. Should include all genes
genome-wide (likely much larger than GRCP_vec). Order does not matter.}
}
\value{
A vector of gene priors to be used in causal gene fine-mapping.
The order of the vector matches the GRCP input.
}
\description{
Estimate gene priors for causal gene fine-mapping.
}
\examples{
data(p1_p12)
data(twas_pi1)
data(grcp)
data(n_cis)
interface_priors(GRCP_vec = grcp, p12 = p1_p12[2],TWAS_pi = twas_pi1,
n_cis = n_cis)
}
