% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isusie_null_wt.R
\name{isusie_null_wt}
\alias{isusie_null_wt}
\title{Estimate null weight for causal gene fine-mapping}
\usage{
isusie_null_wt(p1, p12, n_region_snps)
}
\arguments{
\item{p1}{A prior enrichment parameter from colocalization analysis.
Represents the prior probability that a variant in the region is a
causal GWAS variant but not a causal eQTL.}

\item{p12}{A prior enrichment parameter from colocalization analysis.
Represents the prior probability that a variant in the region is both a
causal GWAS variant and a causal eQTL.}

\item{n_region_snps}{The number of variants in the region of interest for
fine-mapping.}
}
\value{
The null weight (scalar).
}
\description{
Estimate null weight for causal gene fine-mapping
}
\examples{
data(p1_p12)
data(sbams)

isusie_null_wt(p1 = p1_p12[1], p12 = p1_p12[2],n_region_snps=(nrow(sbams)-1))
}
