% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_cis.R
\docType{data}
\name{n_cis}
\alias{n_cis}
\title{Number of cis variants for genes in simulated data.}
\format{
A 39-vector.
}
\usage{
n_cis
}
\value{
A 39-vector.
}
\description{
A numeric 39-vector containing the number of cis variants for each gene in
the simulated data.
}
\examples{
data(n_cis)
}
\keyword{datasets}
