% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_wt.R
\docType{data}
\name{null_wt}
\alias{null_wt}
\title{Null weight for ISuSiE fine-mapping analysis.}
\format{
A scalar.
}
\usage{
null_wt
}
\value{
A scalar.
}
\description{
A scalar null weight to be used as an argument in the isusie function. The
null weight can be derived using the isusie_null_wt function.
}
\examples{
data(null_wt)
}
\keyword{datasets}
