% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p1_p12.R
\docType{data}
\name{p1_p12}
\alias{p1_p12}
\title{Prior enrichment parameters from fastENLOC colocalization analysis for
the simulated data.}
\format{
A 2-vector.
}
\usage{
p1_p12
}
\value{
A 2-vector.
}
\description{
A numeric 2-vector containing the prior enrichment parameters to be used
in gene prior, variant prior, and null weight computations. The first element
is the p1 parameter, and the second is the p12 parameter.
}
\examples{
data(p1_p12)
}
\keyword{datasets}
