% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_expr.R
\docType{data}
\name{pred_expr}
\alias{pred_expr}
\title{Predicted expression levels for genes in the simulated data.}
\format{
A data frame with 706 rows and 39 variables. Each variable
is a gene and each row corresponds to an individual. The row order matches
the column order in the included sbams file.
}
\usage{
pred_expr
}
\value{
A data frame with 706 rows and 39 variables.
}
\description{
A data set containing predicted expression levels for the 39 genes in the
simulated data. We use PTWAS prediction models and a censoring threshold of
CPIP > 0.25 to perform expression prediction. This data is included to
demonstrate the isusie function usage.
}
\examples{
data(pred_expr)
}
\keyword{datasets}
