% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{print.enum}
\alias{print.enum}
\title{Print method for enum objects}
\usage{
\method{print}{enum}(x, ...)
}
\arguments{
\item{x}{An enum object}

\item{...}{Additional arguments (not used)}
}
\value{
No return value, called for side effects.
        Prints a string representation of the enum object to the console.
}
\description{
Prints a human-readable representation of an enum object.
}
