% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{print.enum_generator}
\alias{print.enum_generator}
\title{Print method for enum generators}
\usage{
\method{print}{enum_generator}(x, ...)
}
\arguments{
\item{x}{An enum generator function}

\item{...}{Additional arguments (not used)}
}
\description{
Prints a human-readable representation of an enum generator, showing all possible values.
}
\examples{
Colors <- enum("red", "green", "blue")
print(Colors)  # Output: Enum generator: red, green, blue
}
