% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.Interface}
\alias{print.Interface}
\title{Print method for Interface objects}
\usage{
\method{print}{Interface}(x, ...)
}
\arguments{
\item{x}{The Interface object to print}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the object
}
\description{
Print method for Interface objects
}
