% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print.interface_object}
\alias{print.interface_object}
\title{Print method for interface objects}
\usage{
\method{print}{interface_object}(x, ...)
}
\arguments{
\item{x}{An object implementing an interface}

\item{...}{Additional arguments (not used)}
}
\value{
No return value, called for side effects.
        Prints a human-readable representation of the interface object to the console.
}
\description{
Print method for interface objects
}
