% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.frame.R
\name{rbind.typed_frame}
\alias{rbind.typed_frame}
\title{Combine typed data frames row-wise}
\usage{
\method{rbind}{typed_frame}(..., deparse.level = 1)
}
\arguments{
\item{...}{Typed data frames to combine.}

\item{deparse.level}{See \code{\link[base]{rbind}}.}
}
\value{
The combined typed data frame. The returned object is of class 'typed_frame'
        and inherits all properties (column types, validation rules, etc.) from the first data frame in the list.
}
\description{
This function combines multiple typed data frames row-wise, ensuring type consistency and applying row validation rules.
It extends the base \code{\link[base]{rbind}} function by adding type checks and row validation based on the specified rules for typed data frames.
}
\details{
This version of \code{rbind} for \code{typed_frame} performs extra type checking and row validation to ensure consistency and adherence to specified rules. 
Refer to the base \code{rbind} documentation for additional details on combining data frames: \code{\link[base]{rbind}}.
}
