% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbams.R
\docType{data}
\name{sbams}
\alias{sbams}
\title{sbams file from the simulated data.}
\format{
A data frame with 26987 rows and 709 variables.
}
\usage{
sbams
}
\value{
A data frame with 26987 rows and 709 variables.
}
\description{
A data set containing genotypes and GWAS trait measurements for a
simulated region dataset. The data represents 26986 variants for 706
individuals. The first row contains the GWAS trait measurements,
and rows 2-26987 contain the genotype matrix. The order of columns 4-709
matches the row order of pred_expr.
}
\examples{
data(sbams)
}
\keyword{datasets}
