% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.frame.R
\name{[<-.typed_frame}
\alias{[<-.typed_frame}
\title{Modify a typed data frame using [ ]}
\usage{
\method{[}{typed_frame}(x, i, j) <- value
}
\arguments{
\item{x}{A typed data frame.}

\item{i}{Row index.}

\item{j}{Column index or name.}

\item{value}{The new value to assign.}
}
\value{
The modified typed data frame.
}
\description{
Allows modifying a typed data frame using the [ ] operator, with validation checks.
}
