% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.frame.R
\name{summary.typed_frame}
\alias{summary.typed_frame}
\title{Summary method for typed data frames}
\usage{
\method{summary}{typed_frame}(x)
}
\arguments{
\item{x}{A typed data frame.}
}
\value{
Summary information of the typed data frame.
}
\description{
Provides a summary of the typed data frame, including validation status.
}
