% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas_pi1.R
\docType{data}
\name{twas_pi1}
\alias{twas_pi1}
\title{TWAS prior estimate.}
\format{
A scalar.
}
\usage{
twas_pi1
}
\value{
A scalar.
}
\description{
A scalar TWAS prior estimate to be used in the gene prior computation.
The TWAS prior can be computed using the pi1_fun function.
}
\examples{
data(twas_pi1)
}
\keyword{datasets}
