% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas_z.R
\docType{data}
\name{twas_z}
\alias{twas_z}
\title{TWAS z-score test statistics for simulated genes.}
\format{
A 23400-vector.
}
\usage{
twas_z
}
\value{
A 23400-vector.
}
\description{
A numeric 23400-vector containing the TWAS test statistic for a simulated
dataset. We use this data to demonstrate the pi1_fun function. This data
represents all genes in the full simulated dataset from which the sbams and
pred_expr data are derived. We only include one sbams and pred_expr file in
this package for the purpose of demonstration, but we recommend using the
full transcriptome to estimate pi1 if possible.
}
\examples{
data(twas_z)
}
\keyword{datasets}
