% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_object}
\alias{validate_object}
\title{Validate an object against an interface}
\usage{
validate_object(obj, interface)
}
\arguments{
\item{obj}{The object to validate}

\item{interface}{The interface to validate against}
}
\value{
TRUE if the object is valid, otherwise throws an error
}
\description{
Validate an object against an interface
}
\keyword{internal}
