% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.frame.R
\name{wrap_fun_in_all}
\alias{wrap_fun_in_all}
\title{Modify a user-defined function to return a single logical value}
\usage{
wrap_fun_in_all(user_fun)
}
\arguments{
\item{user_fun}{A user-defined function.}
}
\value{
The modified function.
}
\description{
Modifies a user-defined function to wrap its body in an all() call, ensuring that it returns a single logical value instead of a vector.

It uses bquote() to create a new body for the function.
The .() inside bquote() inserts the original body of the function.
The all() function wraps around the original body.
}
