% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{$.interface_object}
\alias{$.interface_object}
\title{Get a property from an interface object}
\usage{
\method{$}{interface_object}(x, name)
}
\arguments{
\item{x}{An interface object}

\item{name}{The name of the property to get}
}
\value{
The value of the specified property. The class of the returned value
        depends on the property's type as defined in the interface.
}
\description{
Get a property from an interface object
}
