% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{==.enum}
\alias{==.enum}
\title{Equality comparison for enum objects}
\usage{
\method{==}{enum}(e1, e2)
}
\arguments{
\item{e1}{First enum object}

\item{e2}{Second enum object or a character value}
}
\value{
Logical value indicating whether the two objects are equal
}
\description{
Compares two enum objects or an enum object with a character value.
}
