% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{interface}
\alias{interface}
\title{Create an Interface}
\usage{
interface(interface_name, ..., validate_on_access = FALSE)
}
\arguments{
\item{interface_name}{A character string naming the interface}

\item{...}{Property definitions for the interface}

\item{validate_on_access}{Logical, whether to validate on access by default}
}
\value{
An Interface object
}
\description{
Create an Interface
}
