% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{Interface}
\alias{Interface}
\title{Internal function to create an Interface object}
\usage{
Interface(interface_name, properties, validate_on_access = FALSE)
}
\arguments{
\item{interface_name}{A character string naming the interface}

\item{properties}{A list of property definitions}

\item{validate_on_access}{Logical, whether to validate on access by default}
}
\value{
An Interface object
}
\description{
Internal function to create an Interface object
}
\keyword{internal}
