% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{interface}
\alias{interface}
\title{Perform probabilistic causal gene fine-mapping with INTERFACE.}
\usage{
interface(Y, PE, G, p_gene, p_snp, L = 10, null_wt)
}
\arguments{
\item{Y}{An n-vector of GWAS trait measurements for n individuals.}

\item{PE}{A (n x k) dataframe with predicted expression levels for n
individuals and k genes. The order of rows should match Y. Column names
should uniquely identify the corresponding gene.}

\item{G}{A (n x p) dataframe with measured genotypes for n individuals at
k variants. The order of rows should match Y. Column names should uniquely
identify the corresponding variant.}

\item{p_gene}{A k-vector of gene priors. We recommend using the
interface_priors function to estimate these. The order should match the
columns of PE.}

\item{p_snp}{A scalar variant prior. We recommend using the fastENLOC p1
parameter for this quantity.}

\item{L}{An upper bound for the number of causal genes or variants in the
region of interest.}

\item{null_wt}{A scalar for the null weight. We recommend estimating this
parameter using the interface_null_wt function.}
}
\value{
A four-column data frame containing the variable (variant or gene),
PIP, posterior effect estimate, and effect estimate posterior standard
deviation.
}
\description{
Perform probabilistic causal gene fine-mapping with INTERFACE.
}
\examples{
data(sbams)
data(pred_expr)
data(gene_priors)
data(p1_p12)
data(null_wt)

snp <- sbams$V2[2:nrow(sbams)]
G <- t(sbams[2:nrow(sbams),4:ncol(sbams)])
colnames(G) <- snp
Y = as.numeric(sbams[1,4:709])
interface(Y = Y, PE = pred_expr, G = G, p_gene = gene_priors, p_snp = p1_p12[1],null_wt = null_wt)
}
